+++ src/Origins/Reviewers/ReviewerInterface.php
@@ -2,7 +2,10 @@
 
 declare(strict_types=1);
 
-namespace PhpCfdi\SatCatalogosPopulate\Origins;
+namespace PhpCfdi\SatCatalogosPopulate\Origins\Reviewers;
+
+use PhpCfdi\SatCatalogosPopulate\Origins\OriginInterface;
+use PhpCfdi\SatCatalogosPopulate\Origins\Review;
 
 interface ReviewerInterface
 {
