+++ src/Origins/Reviewers/ScrapingReviewer.php
@@ -2,9 +2,16 @@
 
 declare(strict_types=1);
 
-namespace PhpCfdi\SatCatalogosPopulate\Origins;
+namespace PhpCfdi\SatCatalogosPopulate\Origins\Reviewers;
 
 use LogicException;
+use PhpCfdi\SatCatalogosPopulate\Origins\OriginInterface;
+use PhpCfdi\SatCatalogosPopulate\Origins\ResourcesGatewayInterface;
+use PhpCfdi\SatCatalogosPopulate\Origins\Review;
+use PhpCfdi\SatCatalogosPopulate\Origins\ReviewStatus;
+use PhpCfdi\SatCatalogosPopulate\Origins\ScrapingOrigin;
+use PhpCfdi\SatCatalogosPopulate\Origins\ScrapingReviewerLinkExtractor;
+use PhpCfdi\SatCatalogosPopulate\Origins\UrlResponse;
 use RuntimeException;
 
 class ScrapingReviewer implements ReviewerInterface
