+++ tests/Features/Origins/Reviewers/ReviewersTest.php
@@ -2,12 +2,12 @@
 
 declare(strict_types=1);
 
-namespace PhpCfdi\SatCatalogosPopulate\Tests\Features\Origins;
+namespace PhpCfdi\SatCatalogosPopulate\Tests\Features\Origins\Reviewers;
 
 use DateTimeImmutable;
 use PhpCfdi\SatCatalogosPopulate\Origins\ConstantOrigin;
 use PhpCfdi\SatCatalogosPopulate\Origins\Origins;
-use PhpCfdi\SatCatalogosPopulate\Origins\Reviewers;
+use PhpCfdi\SatCatalogosPopulate\Origins\Reviewers\Reviewers;
 use PhpCfdi\SatCatalogosPopulate\Tests\Fixtures\Origins\FakeGateway;
 use PhpCfdi\SatCatalogosPopulate\Tests\TestCase;
